Move fitness files to their own dir
